+++ R/dispersion.R
@@ -47,7 +47,7 @@ setGeneric("dispersion", function(.Object, ...){standardGeneric("dispersion")})
 #' @rdname dispersion-method
 #' @aliases dispersion,slice-method
 #' @name dispersion
-setMethod("dispersion", "slice", function(.Object, query, s_attribute, cqp = FALSE, p_attribute = getOption("polmineR.p_attribute"), freq = FALSE, mc = FALSE, progress = TRUE, verbose = FALSE, ...){
+setMethod("dispersion", "slice", function(.Object, query, s_attribute, cqp = FALSE, p_attribute = getOption("polmineR.p_attribute"), freq = FALSE, mc = FALSE, progress = FALSE, verbose = FALSE, ...){
   dot_list <- list(...)
   if ("sAttribute" %in% names(dot_list)) s_attribute <- dot_list[["sAttribute"]]
   if ("pAttribute" %in% names(dot_list)) p_attribute <- dot_list[["pAttribute"]]
@@ -57,7 +57,7 @@ setMethod("dispersion", "slice", function(.Object, query, s_attribute, cqp = FAL
     s_attribute = s_attribute, p_attribute = p_attribute, freq = freq,
     mc = mc, verbose = verbose, progress = progress
   )
-  dispersion(h, s_attribute = s_attribute, freq = freq)
+  dispersion(h, s_attribute = s_attribute, source = .Object, freq = freq)
 })
 
 #' @rdname dispersion-method
@@ -76,7 +76,7 @@ setMethod("dispersion", "subcorpus", function(
 
 
 #' @rdname dispersion-method
-setMethod("dispersion", "corpus", function(.Object, query, s_attribute, cqp = is.cqp, p_attribute = getOption("polmineR.p_attribute"), freq = FALSE, mc = FALSE, progress = TRUE, verbose = TRUE, ...){
+setMethod("dispersion", "corpus", function(.Object, query, s_attribute, cqp = is.cqp, p_attribute = getOption("polmineR.p_attribute"), freq = FALSE, mc = FALSE, progress = FALSE, verbose = FALSE, ...){
   dot_list <- list(...)
   if ("sAttribute" %in% names(dot_list)) s_attribute <- dot_list[["sAttribute"]]
   if ("pAttribute" %in% names(dot_list)) p_attribute <- dot_list[["pAttribute"]]
@@ -85,27 +85,7 @@ setMethod("dispersion", "corpus", function(.Object, query, s_attribute, cqp = is
     .Object, query = query, cqp = cqp, s_attribute = s_attribute, p_attribute = p_attribute, freq = freq,
     mc = mc, verbose = verbose, progress = progress
   )
-  y <- dispersion(h, s_attribute = s_attribute, freq = freq)
-  
-  if (length(s_attribute) == 1L){
-    if (!freq){
-      s_attr_values <- s_attributes(.Object, s_attribute = s_attribute, unique = TRUE)
-      y <- y[do.call(data.table, setNames(list(s_attr_values, s_attribute), c(s_attribute, "key"))), on = s_attribute]
-      y[, "count" := ifelse(is.na(y[["count"]]), 0L, y[["count"]])]
-    }
-    if (any(is.na(y[["query"]]))) y[, "query" := unique(y[["query"]][!is.na(y[["query"]])])]
-  } else if (length(s_attribute) == 2L){
-    if (!freq){
-      s_attr_values <- s_attributes(.Object, s_attribute = s_attribute[1], unique = TRUE)
-      y <- y[do.call(data.table, setNames(list(s_attr_values, s_attribute[1]), c(s_attribute[1], "key"))), on = s_attribute[1]]
-      y[is.na(y)] <- 0L
-      
-      s_attr_values <- s_attributes(.Object, s_attribute = s_attribute[2], unique = TRUE)
-      s_attr_values_missing <- s_attr_values[!s_attr_values %in% colnames(y)]
-      for (s_attr in s_attr_values_missing) y[, (s_attr) := 0L]
-    }
-  }
-  y
+  dispersion(h, s_attribute = s_attribute, source = .Object, freq = freq)
 })
 
 
@@ -127,7 +107,10 @@ setMethod("dispersion", "character", function(.Object, query, s_attribute, cqp =
 
 
 #' @rdname dispersion-method
-setMethod("dispersion", "hits", function(.Object, s_attribute, freq = FALSE, verbose = TRUE, ...){
+#' @param source The source of the evaluation the hits reported in
+#'   \code{.Object} are based on, a \code{corpus}, \code{subcorpus} or
+#'   \code{partition} object.
+setMethod("dispersion", "hits", function(.Object, source, s_attribute, freq = FALSE, verbose = TRUE, ...){
   
   dot_list <- list(...)
   if ("sAttribute" %in% names(dot_list)) s_attribute <- dot_list[["sAttribute"]]
@@ -149,7 +132,16 @@ setMethod("dispersion", "hits", function(.Object, s_attribute, freq = FALSE, ver
   }
   
   
-  if (length(s_attribute) == 2L){
+  if (length(s_attribute) == 1L){
+    dt <- .Object@stat
+    # ensure that zero matches are reported for all values of the s-attribute
+    if (!freq){
+      s_attr_values <- s_attributes(source, s_attribute = s_attribute, unique = TRUE)
+      dt <- dt[do.call(data.table, setNames(list(s_attr_values, s_attribute), c(s_attribute, "key"))), on = s_attribute]
+      dt[, "count" := ifelse(is.na(dt[["count"]]), 0L, dt[["count"]])]
+    }
+    if (any(is.na(dt[["query"]]))) dt[, "query" := unique(dt[["query"]][!is.na(dt[["query"]])])]
+  } else if (length(s_attribute) == 2L){
     for (s_attr in s_attribute) if ("" %in% .Object@stat[[s_attr]]){
       warning(
         "There is a zero-length character vector for s_attribute ",
@@ -157,14 +149,21 @@ setMethod("dispersion", "hits", function(.Object, s_attribute, freq = FALSE, ver
         ", this will result in a column V1 (V2, V3, ...)."
       )
     }
-    retval <- data.table::dcast.data.table(
+    dt <- data.table::dcast.data.table(
       .Object@stat, formula(paste(s_attribute, collapse = "~")),
       value.var = if (freq) "freq" else "count", fun.aggregate = sum, fill = 0L
-      )  
-  } else if (length(s_attribute) == 1L){
-    retval <- .Object@stat
+    )
+    if (!freq){
+      s_attr_values <- s_attributes(source, s_attribute = s_attribute[1], unique = TRUE)
+      dt <- dt[do.call(data.table, setNames(list(s_attr_values, s_attribute[1]), c(s_attribute[1], "key"))), on = s_attribute[1]]
+      dt[is.na(dt)] <- 0L
+      
+      s_attr_values <- s_attributes(source, s_attribute = s_attribute[2], unique = TRUE)
+      s_attr_values_missing <- s_attr_values[!s_attr_values %in% colnames(dt)]
+      for (s_attr in s_attr_values_missing) dt[, (s_attr) := 0L]
+    }
   } else {
     warning("length(s_attribute) needs to be 1 or 2")
   }
-  retval
+  dt
 })